% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_routine.R
\name{as_indicator}
\alias{as_indicator}
\title{Convert Vectors to Indicators}
\usage{
as_indicator(x, convert_na = FALSE, ...)
}
\arguments{
\item{x}{Vector to be converted}

\item{convert_na}{Whether NAs should be converted to FALSE.}

\item{...}{Specific arguments passed to methods.}
}
\description{
Convert Vectors to Indicators
}
