% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_missing.R
\name{check_missing}
\alias{check_missing}
\alias{check_missing_}
\title{Check missing rate in variables.}
\usage{
check_missing(data, ..., ret_prop = TRUE)

check_missing_(data, ..., .dots, ret_prop = TRUE)
}
\arguments{
\item{data}{A \code{\link[base]{data.frame}} or \code{\link[dplyr]{tbl}}.}

\item{...}{One or more unquoted expressions separated by commas.
You can treat variable names like they are positions, so you can
use expressions like \code{x:y} to select ranges of variables.

Positive values select variables; negative values drop variables.
If the first expression is negative, \code{select()} will automatically
start with all variables.

Use named arguments, e.g. \code{new_name = old_name}, to rename selected variables.

The arguments in \code{...} are automatically \link[rlang:quo]{quoted} and
\link[rlang:eval_tidy]{evaluated} in a context where column names
represent column positions. They also support
\link[rlang:quasiquotation]{unquoting} and splicing. See
\code{vignette("programming")} for an introduction to these concepts.

See \link[tidyselect:select_helpers]{select helpers} for more details and
examples about tidyselect helpers such as \code{starts_with()}, \code{everything()}, ...}

\item{ret_prop}{Whether to return the rate of missing (default) or counts.}

\item{.dots}{Used in conjunction with \code{...} to support both explicit
and implicit arguments.}
}
\description{
Check missing (\code{NA}) proportion or counts of variables. This function
works like \code{\link[dplyr]{summarize_at}} where the missing rate or
count for the selected columns are returned.
}
\section{Functions}{
\itemize{
\item \code{check_missing_}: SE version of check_missing.
}}

\author{
Min Ma
}
