% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_routine.R
\name{tidy_routine}
\alias{tidy_routine}
\alias{tidy_routine.data.frame}
\title{Reshape data into a tidy format.}
\usage{
tidy_routine(data, into, regex = "([[:alnum:]]+)", spread = "var")

\method{tidy_routine}{data.frame}(data, into, regex = "([[:alnum:]]+)",
  spread = "var")
}
\arguments{
\item{data}{A data frame.}

\item{into}{Character vector of variable names to create. \code{into} must
contain the value of \code{spread} and match the matching groups in
\code{regex}.}

\item{regex}{a regular expression used to extract the desired values.}

\item{spread}{The variable used to spread, the default is \code{"var"}.}
}
\description{
Reshape data into a tidy format.
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: base method that deals with the calculation.
}}

\examples{
wide_table <- iris
wide_table$id <- 1:nrow(iris)
tidy_routine(wide_table, into = c("part", "var"),
             regex = "(Sepal|Petal)\\\\.(Length|Width)")
}
