% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_missing.R
\name{check_missing}
\alias{check_missing}
\alias{check_missing.data.frame}
\alias{check_missing.default}
\title{Check missing rate in variables.}
\usage{
check_missing(data, ..., ret_prop = TRUE)

\method{check_missing}{data.frame}(data, ..., ret_prop = TRUE)

\method{check_missing}{default}(data, ...)
}
\arguments{
\item{data}{A \code{\link[base]{data.frame}} or \code{\link[dplyr]{tbl}}.}

\item{...}{One or more unquoted expressions separated by commas.
You can treat variable names like they are positions.

Positive values select variables; negative values to drop variables.
If the first expression is negative, \code{select()} will automatically
start with all variables.

Use named arguments to rename selected variables.

These arguments are automatically \link[rlang:quo]{quoted} and
\link[rlang:eval_tidy]{evaluated} in a context where column names
represent column positions. They support
\link[rlang:quasiquotation]{unquoting} and splicing. See
\code{vignette("programming")} for an introduction to these concepts.}

\item{ret_prop}{Whether to return the rate of missing (default) or counts.}
}
\description{
Check missing (\code{NA}) proportion or counts of variables. This function
works like \code{\link[dplyr]{summarize_at}} where the missing rate or
count for the selected columns are returned.
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: S3 method for data.frame.

\item \code{default}: default method for new NSE generic that dispatches
to the NSE generic for backward compatibility.
}}

\author{
Min Ma
}
