% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_missing.R
\name{check_missing}
\alias{check_missing}
\alias{check_missing_}
\title{Check missing rate in variables.}
\usage{
check_missing(data, ..., ret_prop = TRUE)

check_missing_(data, ..., .dots, ret_prop = TRUE)
}
\arguments{
\item{data}{A \code{\link[base]{data.frame}} or \code{\link[dplyr]{tbl}}.}

\item{...}{Comma separated list of unquoted expressions. You can treat
variable names like they are positions. Use positive values to select
variables; use negative values to drop variables.}

\item{ret_prop}{Whether to return the rate of missing (default) or counts.}

\item{.dots}{Used in conjunction with \code{...} to support both explicit
and implicit arguments.}
}
\description{
Check missing (\code{NA}) proportion or counts of variables. This function
works like \code{\link[dplyr]{summarize_at}} where the missing rate or
count for the selected columns are returned.
}
\section{Functions}{
\itemize{
\item \code{check_missing_}: SE version of check_missing.
}}

\author{
Min Ma
}
