% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_routine.R
\name{ind_to_char}
\alias{ind_to_char}
\alias{ind_to_char.data.frame}
\alias{ind_to_char.grouped_df}
\title{Convert indicator data.frame to character/factor.}
\usage{
ind_to_char(data, col, ..., na_as_false = FALSE, ret_factor = FALSE,
  remove = TRUE)

\method{ind_to_char}{data.frame}(data, col, ..., na_as_false = FALSE,
  ret_factor = FALSE, remove = TRUE)

\method{ind_to_char}{grouped_df}(data, col, ..., na_as_false = FALSE,
  ret_factor = FALSE, remove = TRUE)
}
\arguments{
\item{data}{A \code{\link[base]{data.frame}} or \code{\link[dplyr]{tbl}}.}

\item{col}{Name of the generated column. Use a bare name when using NSE
functions and a character (quoted) name when using SE functions (functions
that end with underscores).}

\item{...}{Specification of indicator columns. Use bare variable names.
Select all variables between \code{x} and \code{z} with \code{x:z}. For more
options, see the \code{\link[dplyr]{select}} documentation.}

\item{na_as_false}{Treat NAs as FALSE in indicators columns.}

\item{ret_factor}{Whether to convert the column into factor.
exhaustive.}

\item{remove}{If \code{TRUE}, remove input column from output data frame.}
}
\description{
This is the reverse operation of using \code{\link[stats]{model.matrix}} on a
factor. \code{ind_to_char} works like \code{tidyr::unite}, it combines
multiple indicator columns into one character/factor column and add it to
the data.
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: Method for data.frame.

\item \code{grouped_df}: Method for grouped_df.
}}

\examples{
# Truth table used by ind_to_char
truth_table <- expand.grid(P = c(TRUE, FALSE, NA), Q = c(TRUE, FALSE, NA))
truth_table$IND <- c("Error", "Q", NA, "P", NA, NA, NA, NA, NA)
truth_table$IND_na_as_false <-
  c("Error", "Q", "Q", "P", NA, NA, "P", NA, NA)
truth_table


# Supports converting the following atomic types to indicator
df <- data.frame(integer_ind = c(2L, 0L, 0L, 0L, 0L, 0L, 0L),
                 # non-zero integer is 1, otherwise 0.
                 logical_ind = c(FALSE, TRUE, FALSE, FALSE, FALSE, FALSE, FALSE),
                 # TRUE is 1.
                 double_ind = c(0, 0, 2.0, 0, 0, 0, 0),
                 # non-zero double is 1.
                 char_ind = c("FALSE", "FALSE", "F", "TRUE", "T", "F", "F"),
                 # "T" and "TRUE" converts to 1.
                 factor_ind = factor(c(1, 1, 1, 1, 1, 0, 1), levels = c(0, 1),
                                     labels = c(TRUE, FALSE)),
                 # Factors are converted on levels, see as.logical.factor()
                 # Note that here 0 = TRUE, 1 = FALSE, but 0 is converted to
                 # logical TRUE.
                 stringsAsFactors = FALSE)

# Keep original columns with remove = FALSE
ind_to_char(df, col = new_y, integer_ind, logical_ind, double_ind,
            char_ind, factor_ind, remove = FALSE)



# ind_to_char as complement to use model.matrix on a factor
model_df <- data.frame(x = 1:6, y = factor(c(letters[1:5], NA)))
model_df
ind_df <- as.data.frame(model.matrix(~ x + y - 1,
                                     model.frame(model_df,
                                                 na.action = na.pass)))
ind_df  # an indicator matrix with NAs

# Convert back to one column like model_df
ind_to_char(ind_df, new_y, ya, yb, yc, yd, ye)

# Returns a factor column
ind_to_char(ind_df, col = new_y, ya:ye, ret_factor = TRUE)


# Use select helper functions as in dplyr::select
ind_to_char(ind_df, new_y, contains("y"), remove = FALSE)
ind_to_char(ind_df, new_y, starts_with("y"), remove = FALSE)

# Using SE
ind_to_char(ind_df, col = "new_y", "ya", "yb", "yc", "yd", "ye")
}
